Crate cipha_cli

cipha-cli

cipha-cli is a CLI that supports various classifical ciphers such as ROT13, Caesar, Vigenere, Gematria and more. It is an interface to the cipha library or crate.

Usage

cipha-cli encode --cipher rot13 --message "Hello, World!"
cipha-cli encode --cipher caesar --message "Hello, World!" --shift 3

Structs

Cli 🔒

A simple CLI for ciphers and cryptography.

Enums

Command 🔒

Subcommands for the CLI.

Functions

decode_message 🔒

Decodes a message using the specified cipher.

encode_message 🔒

Encodes a message using the specified cipher.

get_message 🔒

Retrieves the message from either the command-line argument or a file.

main 🔒